requirements.txt
fastapi
uvicorn[standard]
httpx
beautifulsoup4
lxml
sqlalchemy
aiosqlite
python-dotenv
